chore(ci_vis): make civis integrations internal
release note

fmt

fix more imports
