+++ ddtrace/contrib/internal/pytest/_benchmark_utils.py
@@ -1,7 +1,7 @@
 import pytest
 
-from ddtrace.contrib.pytest._utils import _get_test_id_from_item
-from ddtrace.contrib.pytest_benchmark.constants import PLUGIN_METRICS_V2
+from ddtrace.contrib.internal.pytest._utils import _get_test_id_from_item
+from ddtrace.contrib.internal.pytest_benchmark.constants import PLUGIN_METRICS_V2
 from ddtrace.internal.logger import get_logger
 from ddtrace.internal.test_visibility._benchmark_mixin import BenchmarkDurationData
 from ddtrace.internal.test_visibility.api import InternalTest
