+++ ddtrace/contrib/internal/pytest/_retry_utils.py
@@ -8,8 +8,8 @@ from _pytest.logging import caplog_records_key
 from _pytest.runner import CallInfo
 import pytest
 
-from ddtrace.contrib.pytest._types import tmppath_result_key
-from ddtrace.contrib.pytest._utils import _TestOutcome
+from ddtrace.contrib.internal.pytest._types import tmppath_result_key
+from ddtrace.contrib.internal.pytest._utils import _TestOutcome
 from ddtrace.ext.test_visibility.api import TestExcInfo
 from ddtrace.ext.test_visibility.api import TestStatus
 from ddtrace.internal import core
