+++ ddtrace/contrib/internal/pytest_bdd/constants.py
@@ -1,0 +1,2 @@
+FRAMEWORK = "pytest_bdd"
+STEP_KIND = "pytest_bdd.step"
