+++ ddtrace/contrib/pytest/constants.py
@@ -1,11 +1,14 @@
-FRAMEWORK = "pytest"
-KIND = "test"
+from ddtrace.contrib.internal.pytest.constants import *  # noqa: F403
+from ddtrace.internal.utils.deprecations import DDTraceDeprecationWarning
+from ddtrace.vendor.debtcollector import deprecate
 
 
-# XFail Reason
-XFAIL_REASON = "pytest.xfail.reason"
+def __getattr__(name):
+    deprecate(
+        ("%s.%s is deprecated" % (__name__, name)),
+        category=DDTraceDeprecationWarning,
+    )
 
-ITR_MIN_SUPPORTED_VERSION = (7, 2, 0)
-RETRIES_MIN_SUPPORTED_VERSION = (7, 0, 0)
-EFD_MIN_SUPPORTED_VERSION = RETRIES_MIN_SUPPORTED_VERSION
-ATR_MIN_SUPPORTED_VERSION = RETRIES_MIN_SUPPORTED_VERSION
+    if name in globals():
+        return globals()[name]
+    raise AttributeError("%s has no attribute %s", __name__, name)
