+++ ddtrace/contrib/pytest_bdd/constants.py
@@ -1,2 +1,14 @@
-FRAMEWORK = "pytest_bdd"
-STEP_KIND = "pytest_bdd.step"
+from ddtrace.contrib.internal.pytest_bdd.constants import *  # noqa: F403
+from ddtrace.internal.utils.deprecations import DDTraceDeprecationWarning
+from ddtrace.vendor.debtcollector import deprecate
+
+
+def __getattr__(name):
+    deprecate(
+        ("%s.%s is deprecated" % (__name__, name)),
+        category=DDTraceDeprecationWarning,
+    )
+
+    if name in globals():
+        return globals()[name]
+    raise AttributeError("%s has no attribute %s", __name__, name)
