+++ ddtrace/contrib/pytest_bdd/plugin.py
@@ -1,20 +1,14 @@
-from ddtrace import DDTraceDeprecationWarning
-from ddtrace.contrib.pytest._utils import _USE_PLUGIN_V2
-from ddtrace.contrib.pytest.plugin import is_enabled as is_ddtrace_enabled
+from ddtrace.contrib.internal.pytest_bdd.plugin import *  # noqa: F403
+from ddtrace.internal.utils.deprecations import DDTraceDeprecationWarning
 from ddtrace.vendor.debtcollector import deprecate
 
 
-def pytest_configure(config):
-    if config.pluginmanager.hasplugin("pytest-bdd") and config.pluginmanager.hasplugin("ddtrace"):
-        if not _USE_PLUGIN_V2:
-            if is_ddtrace_enabled(config):
-                from ._plugin import _PytestBddPlugin
+def __getattr__(name):
+    deprecate(
+        ("%s.%s is deprecated" % (__name__, name)),
+        category=DDTraceDeprecationWarning,
+    )
 
-                deprecate(
-                    "the ddtrace.pytest_bdd plugin is deprecated",
-                    message="it will be integrated with the main pytest ddtrace plugin",
-                    removal_version="3.0.0",
-                    category=DDTraceDeprecationWarning,
-                )
-
-                config.pluginmanager.register(_PytestBddPlugin(), "_datadog-pytest-bdd")
+    if name in globals():
+        return globals()[name]
+    raise AttributeError("%s has no attribute %s", __name__, name)
