+++ ddtrace/contrib/unittest/constants.py
@@ -1,8 +1,14 @@
-COMPONENT_VALUE = "unittest"
-FRAMEWORK = "unittest"
-KIND = "test"
+from ddtrace.contrib.internal.unittest.constants import *  # noqa: F403
+from ddtrace.internal.utils.deprecations import DDTraceDeprecationWarning
+from ddtrace.vendor.debtcollector import deprecate
 
-TEST_OPERATION_NAME = "unittest.test"
-SUITE_OPERATION_NAME = "unittest.test_suite"
-SESSION_OPERATION_NAME = "unittest.test_session"
-MODULE_OPERATION_NAME = "unittest.test_module"
+
+def __getattr__(name):
+    deprecate(
+        ("%s.%s is deprecated" % (__name__, name)),
+        category=DDTraceDeprecationWarning,
+    )
+
+    if name in globals():
+        return globals()[name]
+    raise AttributeError("%s has no attribute %s", __name__, name)
