+++ pyproject.toml
@@ -55,9 +55,9 @@ ddtrace-run = "ddtrace.commands.ddtrace_run:main"
 ddcontextvars_context = "ddtrace.internal.opentelemetry.context:DDRuntimeContext"
 
 [project.entry-points.pytest11]
-ddtrace = "ddtrace.contrib.pytest.plugin"
-"ddtrace.pytest_bdd" = "ddtrace.contrib.pytest_bdd.plugin"
-"ddtrace.pytest_benchmark" = "ddtrace.contrib.pytest_benchmark.plugin"
+ddtrace = "ddtrace.contrib.internal.pytest.plugin"
+"ddtrace.pytest_bdd" = "ddtrace.contrib.internal.pytest_bdd.plugin"
+"ddtrace.pytest_benchmark" = "ddtrace.contrib.internal.pytest_benchmark.plugin"
 
 [project.entry-points.'ddtrace.products']
 "code-origin-for-spans" = "ddtrace.debugging._products.code_origin.span"
