+++ tests/contrib/pytest_benchmark/test_pytest_benchmark.py
@@ -1,24 +1,24 @@
 import os
 
-from ddtrace.contrib.pytest_benchmark.constants import BENCHMARK_INFO
-from ddtrace.contrib.pytest_benchmark.constants import BENCHMARK_MEAN
-from ddtrace.contrib.pytest_benchmark.constants import BENCHMARK_RUN
-from ddtrace.contrib.pytest_benchmark.constants import STATISTICS_HD15IQR
-from ddtrace.contrib.pytest_benchmark.constants import STATISTICS_IQR
-from ddtrace.contrib.pytest_benchmark.constants import STATISTICS_IQR_OUTLIERS
-from ddtrace.contrib.pytest_benchmark.constants import STATISTICS_LD15IQR
-from ddtrace.contrib.pytest_benchmark.constants import STATISTICS_MAX
-from ddtrace.contrib.pytest_benchmark.constants import STATISTICS_MEAN
-from ddtrace.contrib.pytest_benchmark.constants import STATISTICS_MEDIAN
-from ddtrace.contrib.pytest_benchmark.constants import STATISTICS_MIN
-from ddtrace.contrib.pytest_benchmark.constants import STATISTICS_N
-from ddtrace.contrib.pytest_benchmark.constants import STATISTICS_OPS
-from ddtrace.contrib.pytest_benchmark.constants import STATISTICS_OUTLIERS
-from ddtrace.contrib.pytest_benchmark.constants import STATISTICS_Q1
-from ddtrace.contrib.pytest_benchmark.constants import STATISTICS_Q3
-from ddtrace.contrib.pytest_benchmark.constants import STATISTICS_STDDEV
-from ddtrace.contrib.pytest_benchmark.constants import STATISTICS_STDDEV_OUTLIERS
-from ddtrace.contrib.pytest_benchmark.constants import STATISTICS_TOTAL
+from ddtrace.contrib.internal.pytest_benchmark.constants import BENCHMARK_INFO
+from ddtrace.contrib.internal.pytest_benchmark.constants import BENCHMARK_MEAN
+from ddtrace.contrib.internal.pytest_benchmark.constants import BENCHMARK_RUN
+from ddtrace.contrib.internal.pytest_benchmark.constants import STATISTICS_HD15IQR
+from ddtrace.contrib.internal.pytest_benchmark.constants import STATISTICS_IQR
+from ddtrace.contrib.internal.pytest_benchmark.constants import STATISTICS_IQR_OUTLIERS
+from ddtrace.contrib.internal.pytest_benchmark.constants import STATISTICS_LD15IQR
+from ddtrace.contrib.internal.pytest_benchmark.constants import STATISTICS_MAX
+from ddtrace.contrib.internal.pytest_benchmark.constants import STATISTICS_MEAN
+from ddtrace.contrib.internal.pytest_benchmark.constants import STATISTICS_MEDIAN
+from ddtrace.contrib.internal.pytest_benchmark.constants import STATISTICS_MIN
+from ddtrace.contrib.internal.pytest_benchmark.constants import STATISTICS_N
+from ddtrace.contrib.internal.pytest_benchmark.constants import STATISTICS_OPS
+from ddtrace.contrib.internal.pytest_benchmark.constants import STATISTICS_OUTLIERS
+from ddtrace.contrib.internal.pytest_benchmark.constants import STATISTICS_Q1
+from ddtrace.contrib.internal.pytest_benchmark.constants import STATISTICS_Q3
+from ddtrace.contrib.internal.pytest_benchmark.constants import STATISTICS_STDDEV
+from ddtrace.contrib.internal.pytest_benchmark.constants import STATISTICS_STDDEV_OUTLIERS
+from ddtrace.contrib.internal.pytest_benchmark.constants import STATISTICS_TOTAL
 from ddtrace.ext.test import TEST_TYPE
 from tests.contrib.pytest.test_pytest import PytestTestCaseBase
