+++ tests/contrib/unittest/test_unittest.py
@@ -7,15 +7,15 @@ import pytest
 from ddtrace.constants import ERROR_MSG
 from ddtrace.constants import ERROR_TYPE
 from ddtrace.constants import SPAN_KIND
-from ddtrace.contrib.unittest.constants import COMPONENT_VALUE
-from ddtrace.contrib.unittest.constants import FRAMEWORK
-from ddtrace.contrib.unittest.constants import KIND
-from ddtrace.contrib.unittest.constants import MODULE_OPERATION_NAME
-from ddtrace.contrib.unittest.constants import SESSION_OPERATION_NAME
-from ddtrace.contrib.unittest.constants import SUITE_OPERATION_NAME
-from ddtrace.contrib.unittest.constants import TEST_OPERATION_NAME
-from ddtrace.contrib.unittest.patch import _set_tracer
-from ddtrace.contrib.unittest.patch import patch
+from ddtrace.contrib.internal.unittest.constants import COMPONENT_VALUE
+from ddtrace.contrib.internal.unittest.constants import FRAMEWORK
+from ddtrace.contrib.internal.unittest.constants import KIND
+from ddtrace.contrib.internal.unittest.constants import MODULE_OPERATION_NAME
+from ddtrace.contrib.internal.unittest.constants import SESSION_OPERATION_NAME
+from ddtrace.contrib.internal.unittest.constants import SUITE_OPERATION_NAME
+from ddtrace.contrib.internal.unittest.constants import TEST_OPERATION_NAME
+from ddtrace.contrib.internal.unittest.patch import _set_tracer
+from ddtrace.contrib.internal.unittest.patch import patch
 from ddtrace.ext import SpanTypes
 from ddtrace.ext import test
 from ddtrace.ext.ci import RUNTIME_VERSION
