+++ tests/contrib/unittest/test_unittest_patch.py
@@ -3,12 +3,12 @@
 # removed the ``_generated`` suffix from the file name, to prevent the content
 # from being overwritten by future re-generations.
 
-from ddtrace.contrib.unittest.patch import get_version
-from ddtrace.contrib.unittest.patch import patch
+from ddtrace.contrib.internal.unittest.patch import get_version
+from ddtrace.contrib.internal.unittest.patch import patch
 
 
 try:
-    from ddtrace.contrib.unittest.patch import unpatch
+    from ddtrace.contrib.internal.unittest.patch import unpatch
 except ImportError:
     unpatch = None
 from tests.contrib.patch import PatchTestCase
