+++ tests/internal/test_module.py
@@ -578,17 +578,6 @@ def __getattr__(name):
             "ddtrace.contrib.trace_utils",
             "ddtrace.contrib.trace_utils_async",
             "ddtrace.contrib.trace_utils_redis",
-            # TODO: The following contrib modules are part of the public API (unlike most integrations).
-            # We should consider privatizing the internals of these integrations.
-            "ddtrace.contrib.unittest.patch",
-            "ddtrace.contrib.unittest.constants",
-            "ddtrace.contrib.pytest.constants",
-            "ddtrace.contrib.pytest.newhooks",
-            "ddtrace.contrib.pytest.plugin",
-            "ddtrace.contrib.pytest_benchmark.constants",
-            "ddtrace.contrib.pytest_benchmark.plugin",
-            "ddtrace.contrib.pytest_bdd.constants",
-            "ddtrace.contrib.pytest_bdd.plugin",
         ]
     )
